+++ src/main/kotlin/com/insyncwithfoo/pyright/runner/Notifier.kt
@@ -1,63 +1,50 @@
 package com.insyncwithfoo.pyright.runner
 
-import com.insyncwithfoo.pyright.PyrightIcon
 import com.insyncwithfoo.pyright.PyrightInspection
+import com.insyncwithfoo.pyright.addSimpleExpiringAction
+import com.insyncwithfoo.pyright.openingPyrightNotifications
+import com.insyncwithfoo.pyright.prettify
+import com.insyncwithfoo.pyright.pyrightNotificationGroup
+import com.insyncwithfoo.pyright.runThenNotify
 import com.intellij.notification.Notification
-import com.intellij.notification.NotificationAction
-import com.intellij.notification.NotificationAction.createSimpleExpiring
-import com.intellij.notification.NotificationGroupManager
 import com.intellij.openapi.project.Project
 import com.intellij.profile.codeInspection.InspectionProjectProfileManager
-import java.lang.ref.WeakReference
 
 
-private infix fun Notification.sameAs(other: Notification?) =
-    title == other?.title && content == other.content
+private infix fun Notification.sameAs(other: Notification) =
+    title == other.title && content == other.content
+
+
+private fun Project.disablePyrightInspection() {
+    val inspectionManager = InspectionProjectProfileManager.getInstance(this)
+    val profile = inspectionManager.currentProfile
+    val pyrightInspection = profile.allTools.find { it.tool.shortName == PyrightInspection.SHORT_NAME }
+    
+    pyrightInspection?.isEnabled = false
+    profile.profileChanged()
+}
 
 
 internal class Notifier(private val project: Project) {
     
-    private val groupManager = NotificationGroupManager.getInstance()
-    private val group = groupManager.getNotificationGroup(ID)
+    private val group = pyrightNotificationGroup()
     
     fun notify(exception: PyrightException) {
-        val notification = exception.createNotification(group)
+        val newNotification = exception.createNotification(group)
+        val notified = project.openingPyrightNotifications
         
-        if (notified.any { notification sameAs it.get() }) {
-            return
+        if (!notified.any { it sameAs newNotification }) {
+            notify(newNotification)
         }
-        
-        notified.add(WeakReference(notification))
-        
-        return notify(notification)
     }
     
     private fun notify(notification: Notification) {
-        notification.run {
-            isImportant = false
-            icon = PyrightIcon.COLORED_SMALL
-            addAction(disablePlugin(project))
-            
-            notify(project)
+        notification.runThenNotify(project) {
+            prettify()
+            addSimpleExpiringAction("Disable plugin (project)") {
+                project.disablePyrightInspection()
+            }
         }
-    }
-    
-    private fun disablePlugin(project: Project): NotificationAction {
-        return createSimpleExpiring(DISABLE_PLUGIN_TEXT) {
-            val inspectionManager = InspectionProjectProfileManager.getInstance(project)
-            val profile = inspectionManager.currentProfile
-            val pyrightInspection = profile.allTools.find { it.tool.shortName == PyrightInspection.SHORT_NAME }
-            
-            pyrightInspection?.isEnabled = false
-            profile.profileChanged()
-            notified.clear()
-        }
-    }
-    
-    companion object {
-        private const val ID = "Pyright"
-        private const val DISABLE_PLUGIN_TEXT = "Disable plugin (project)"
-        private val notified = mutableListOf<WeakReference<Notification>>()
     }
     
 }